+++ src/render/draw_custom.ts
@@ -27,6 +27,12 @@ export function drawCustom(painter: Painter, sourceCache: SourceCache, layer: Cu
             define: projection.shaderDefine,
         },
         defaultProjectionData: projectionData,
+        getMatrixForModel(location, altitude) {
+            return transform.getMatrixForModel(location, altitude)
+        },
+        getProjectionData(params) {
+            return transform.getProjectionData(params);
+        },
     };
 
     const renderingMode = implementation.renderingMode ? implementation.renderingMode : '2d';